refactor(ui): extract model context window lookup to utils/models.js
Move MODEL_CONTEXT_WINDOWS table and getContextWindowSize() from
ConversationPropertiesPanel.js to a shared utility module.
Both ChatInput (context %) and ConversationPropertiesPanel now import
from the same source.
